Move more code into host-lib's `TestStand`
